requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cssselect==1.1.0
cssutils==2.3.0
idna==3.2
lxml==4.6.3
premailer==3.10.0
python-dotenv==0.19.0
requests==2.26.0
selenium==3.141.0
urllib3==1.26.6
XlsxWriter==3.0.1
yagmail==0.14.256
